# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PDAM Tiara Material Tracking & Verification System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: string
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete/deactivate a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Branches
...
  /work-orders:
    get:
      operationId: listWorkOrders
      tags: [workOrders]
      summary: List work orders
      parameters:
        - name: branchId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of work orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrder"
    post:
      operationId: createWorkOrder
      tags: [workOrders]
      summary: Create a work order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderInput"
      responses:
        "201":
          description: Work order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
  /work-orders/{id}:
    get:
      operationId: getWorkOrder
      tags: [workOrders]
      summary: Get work order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Work order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"

  # Material Requests
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # Auth
...
    User:
      type: object
      required: [id, name, email, role, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin_gudang, petugas_lapangan, spi, direksi, superadmin]
        branchId:
          type: ["string", "null"]
        branchName:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin_gudang, petugas_lapangan, spi, direksi, superadmin]
        branchId:
          type: ["string", "null"]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin_gudang, petugas_lapangan, spi, direksi, superadmin]
        branchId:
          type: ["string", "null"]
        isActive:
          type: boolean

    # Branch
...
    WorkOrder:
      type: object
      required: [id, woNumber, branchId, createdAt]
      properties:
        id:
          type: string
        woNumber:
          type: string
        description:
          type: ["string", "null"]
        branchId:
          type: string
        branchName:
          type: ["string", "null"]
        createdById:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    WorkOrderInput:
      type: object
      required: [woNumber, branchId]
      properties:
        woNumber:
          type: string
        description:
          type: ["string", "null"]
        branchId:
          type: string

    # Material Request
...